Arduino can now start string pluck
InputDevice.ino now checks the last 3 samples from the pressure button in order to determine a valid press, and sort out noise. It only sends data through serial when there is a valid press.
KarplusStrong.py responds to presence of data from the Arduino, and starts a string pluck. It does not yet use the sent values.